--- a/lib/help.js
+++ b/lib/help.js
@@ -57,7 +57,7 @@ module.exports = function(argv) {
             console.log('');
             console.log('usage: index.js map [--in-network=<FILE.geojson>] [--in-address=<FILE.geojson>] [--output=<FILE.geojson>]');
             console.log('                    [--skip-import] [--debug] [--error-network <FILE>] [--error-address <FILE>]');
-            console.log('                    [--post <cardinality>,...] [--intersections] [--props <prop>,...]');
+            console.log('                    [--post <cardinality>,...] [--intersections] [--props <prop>,...] [--warn <FILE]');
             console.log('');
             console.log('[options]:');
             console.log('   --in-network=<FILE.geojson>     geojson of street network [optional if --skip-import is used]');
@@ -78,6 +78,7 @@ module.exports = function(argv) {
             console.log('   --debug                         [optional] Gives much richer info for `debug` mode module');
             console.log('   --error-network <FILE>          [optional] Output invalid features to a given file');
             console.log('   --error-map <FILE>              [optional] Output invalid features to a given file');
+            console.log('   --warn <FILE>                   [optional] Output build warnings to a file');
             break;
         case ('conflate'):
             console.log('');