--- a/lib/map.js
+++ b/lib/map.js
@@ -32,6 +32,7 @@ function main(argv, cb) {
     if (Array.isArray(argv)) {
         argv = require('minimist')(argv, {
             string: [
+                'warn',
                 'props',
                 'post',
                 'output',
@@ -96,6 +97,10 @@ function main(argv, cb) {
 
     const output = fs.createWriteStream(path.resolve(__dirname, '..', argv.output));
 
+    if (argv.warn) {
+        argv.warn = String(path.resolve(__dirname, '..', argv.warn));
+    }
+
     const poolConf = {
         max: CPUS,
         user: 'postgres',
@@ -124,7 +129,6 @@ function main(argv, cb) {
             })
         };
 
-        console.error('ok - address import');
         console.time('ok - address imported');
         import_addr({
             db: argv.db,
@@ -135,7 +139,6 @@ function main(argv, cb) {
         });
         console.timeEnd('ok - address imported')
 
-        console.error('ok - network import');
         console.time('ok - network imported');
         import_net({
             db: argv.db,
@@ -164,8 +167,6 @@ function main(argv, cb) {
                 index.setMeta('country', argv.country);
             }
 
-            console.error('ok - imported data');
-
             clusterGeom();
         });
     }
@@ -176,15 +177,17 @@ function main(argv, cb) {
     function clusterGeom(err, segs) {
         if (err) return cb(err);
 
+        console.time('ok - network geometries clustered');
         cluster.network((err) => {
             if (err) return cb(err);
 
-            console.error('ok - network geometries clustered');
+            console.timeEnd('ok - network geometries clustered');
 
+            console.time('ok - created cluster indexes');
             cluster.optimize((err) => {
                 if (err) return cb(err);
 
-                console.error('ok - created cluster indexes');
+                console.timeEnd('ok - created cluster indexes');
 
                 return matcher();
             });
@@ -201,6 +204,7 @@ function main(argv, cb) {
             if (!res.rows.length || res.rows[0].max <= 0) return cb(new Error('address has no addresses to match!'));
 
             const addrNum = parseInt(res.rows[0].max);
+            console.time('ok - cross matched data');
             const bar = new Prog('ok - Cross Matching Data [:bar] :percent :etas', {
                 complete: '=',
                 incomplete: ' ',
@@ -237,8 +241,14 @@ function main(argv, cb) {
                         });
 
                         if (!active) {
-                            console.error('ok - addresses assigned to linestring');
-                            return cluster.address(intersection);
+                            console.timeEnd('ok - cross matched data');
+
+                            console.time('ok - clustered addresses');
+                            return cluster.address(() => {
+                                console.timeEnd('ok - clustered addresses');
+
+                                return intersections();
+                            });
                         }
                     } else {
                         let max = addrNum_it = addrNum_it + 5000;
@@ -265,16 +275,14 @@ function main(argv, cb) {
      * Manage intersection generation
      */
     async function intersection() {
-        console.error('ok - orphans output');
-
         try {
             const intsec = new (require('./map/intersections'))({
                 pool: poolConf
             });
 
+            console.time('ok - generated intersections');
             await intsec.generate();
-
-            console.error('ok - generated intersections');
+            console.timeEnd('ok - generated intersections');
 
             return splitter();