--- a/lib/map.js
+++ b/lib/map.js
@@ -247,7 +247,7 @@ function main(argv, cb) {
                             return cluster.address(() => {
                                 console.timeEnd('ok - clustered addresses');
 
-                                return intersections();
+                                return intersection();
                             });
                         }
                     } else {
@@ -363,6 +363,7 @@ function main(argv, cb) {
 
                 opts = {
                     id: id,
+                    warn: argv.warn,
                     props: argv.props,
                     post: argv.post,
                     label: argv.label,
@@ -387,11 +388,19 @@ function main(argv, cb) {
         orphanQ = new Queue();
 
         orphanQ.defer((done) => {
-            orphan.address(done)
+            console.time('ok - output orphaned address');
+            orphan.address(() => {
+                console.timeEnd('ok - output orphaned address');
+                done();
+            });
         });
 
         orphanQ.defer((done) => {
-            orphan.network(done)
+            console.time('ok - output orphaned network');
+            orphan.network(() => {
+                console.timeEnd('ok - output orphaned network');
+                done();
+            });
         });
 
         orphanQ.awaitAll(finalize);
@@ -404,12 +413,10 @@ function main(argv, cb) {
     function finalize(err) {
         if (err) return cb(err);
 
-        console.error('ok - Doing a bit of housecleaning');
-
+        console.time('ok - Doing a bit of housecleaning');
         output.end();
         pool.end();
-
-        console.error('ok - You\'re all set!');
+        console.timeEnd('ok - Doing a bit of housecleaning');
 
         return cb();
     }