--- a/lib/map/cluster.js
+++ b/lib/map/cluster.js
@@ -332,7 +332,6 @@ class Cluster {
 
             COMMIT;
         `, (err, res) => {
-            if (!err) console.error('ok - clustered addresses');
             return cb(err);
         });
     }