--- a/lib/map/cluster.js
+++ b/lib/map/cluster.js
@@ -15,6 +15,13 @@ const numCaseChanges = require('../label/util').numCaseChanges;
 class Cluster {
     constructor(opts) {
         opts = opts || {};
+
+        if (opts.warn) {
+            this.warn = fs.createWriteStream(args.warn, {
+                flags: 'a' 
+            });
+        }
+
         this.pool = opts.pool;
         if (opts.label) {
             this.label = require('../label/' + opts.label)();
@@ -71,7 +78,9 @@ class Cluster {
                 };
             });
         } else if (!segDup.some((dup) => { return dup })) {
-            console.error(`WARN: detected unhandled ClusterDup on network_cluster.id: ${id}`);
+            if (this.warn) {
+                this.warn.write(`WARN: detected unhandled ClusterDup on network_cluster.id: ${id}\n`);
+            }
         }
 
         let orphanSegs = []; //Segs that needs to be matched based on prox. at the end with segments in newSegs
@@ -290,7 +299,6 @@ class Cluster {
                     GROUP BY names
                 ) addr;
         `, (err, res) => {
-            if (!err) console.error('ok - clustered orphan addresses');
             return cb(err);
         });
     }