--- a/lib/map/cluster.js
+++ b/lib/map/cluster.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 const Cursor = require('pg-cursor');
 const turf = require('@turf/turf');
 const Queue = require('d3-queue').queue;