--- a/lib/map/post.js
+++ b/lib/map/post.js
@@ -29,8 +29,10 @@ class Post {
 
         this.opts.post = this.opts.post || [];
 
-        if (!this.opts.label && !this.opts.tokens) {
-            console.error('WARN: map.orphanAddr() using titlecase behavior on non-English data; English capitalization rules will be applied, potentially in error');
+        if (args.warn) {
+            this.opts.warn = fs.createWriteStream(args.warn, {
+                flags: 'a'
+            });
         }
 
         this.opts.label = require('../label/' + (opts.label || 'titlecase'))()