--- a/lib/map/post.js
+++ b/lib/map/post.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 const defaults = [ //Mandatory Steps
     require('../post/intersections').post,
     require('../post/props').post,