--- a/lib/post/text.js
+++ b/lib/post/text.js
@@ -13,7 +13,10 @@ function post(feat, opts = {}) {
             feat.properties[k] = opts.label(feat.properties[k])
 
             if (feat.properties[k].split(',').length > 10) {
-                console.error(`ok - WARN too many synonyms - truncating!: {}`, feat.properties[k]);
+                if (opts.warn) {
+                    opts.warn.write(`WARN: too many synonyms - truncating!: ${feat.properties[k]}\n`);
+                }
+
                 feat.properties[k] = feat.properties[k].split(',').splice(0, 10).join(',');
             }
         });