Output time for each processing step

### Map Acceptance Test

```
ok - address imported: 464.605ms                                                              
ok - network imported: 1860.033ms                                     
ok - network geometries clustered: 1242.723ms                                                           
ok - created cluster indexes: 289.484ms      
ok - Cross Matching Data [====================] 100% 0.0s                                               
ok - cross matched data: 108532.830ms                         
ok - clustered addresses: 67.486ms                                                                      
ok - generated intersections: 12882.181ms                                                               
ok - Splitting Data [====================] 100% 0.0s
ok - output orphaned address: 52.985ms                                      
ok - output orphaned network: 12716.435ms
ok - Doing a bit of housecleaning: 0.152ms
```