@ingalls 2019-02-25
### US_DC

```
ok - address imported: 8504.499ms
ok - network imported: 1009.327ms
ok - network geometries clustered: 967.906ms
ok - created cluster indexes: 2566.034ms
ok - Cross Matching Data [====================] 100% 0.0s
ok - cross matched data: 187692.144ms
ok - clustered addresses: 1015.545ms
ok - generated intersections: 6449.911ms
ok - Splitting Data [====================] 100% 0.0s
ok - output orphaned address: 131.467ms
ok - output orphaned network: 1575.318ms
ok - Doing a bit of housecleaning: 0.357ms
```